+++ stage3-gnome.sh
@@ -126,15 +126,15 @@ install -t /usr/share/licenses/libcloudproviders -Dm644 ../LICENSE
 cd ../..
 rm -rf libcloudproviders-0.3.1
 # gtk4
-tar -xf gtk-4.6.6.tar.xz
-cd gtk-4.6.6
+tar -xf gtk-4.6.7.tar.xz
+cd gtk-4.6.7
 mkdir build; cd build
 meson --prefix=/usr --buildtype=minsize -Dx11-backend=true -Dwayland-backend=true -Dbroadway-backend=true -Dmedia-ffmpeg=enabled -Ddemos=false -Dmedia-gstreamer=enabled -Dprint-cups=enabled -Dvulkan=enabled -Dcloudproviders=enabled -Dsysprof=enabled -Dtracker=enabled -Dcolord=enabled -Dgtk_doc=false -Dman-pages=true -Dbuild-examples=false -Dbuild-tests=false -Dinstall-tests=false ..
 ninja
 ninja install
 install -t /usr/share/licenses/gtk4 -Dm644 ../COPYING
 cd ../..
-rm -rf gtk-4.6.6
+rm -rf gtk-4.6.7
 # JS91.
 tar -xf firefox-91.12.0esr.source.tar.xz
 cd firefox-91.12.0
@@ -183,25 +183,25 @@ install -t /usr/share/licenses/gnome-video-effects -Dm644 ../COPYING
 cd ../..
 rm -rf gnome-video-effects-0.5.0
 # Grilo
-tar -xf grilo-0.3.14.tar.xz
-cd grilo-0.3.14
+tar -xf grilo-0.3.15.tar.xz
+cd grilo-0.3.15
 mkdir build; cd build
 meson --prefix=/usr --buildtype=minsize -Denable-gtk-doc=false -Denable-test-ui=false ..
 ninja
 ninja install
 install -t /usr/share/licenses/grilo -Dm644 ../COPYING
 cd ../..
-rm -rf grilo-0.3.14
+rm -rf grilo-0.3.15
 # grilo-plugins
-tar -xf grilo-plugins-0.3.14.tar.xz
-cd grilo-plugins-0.3.14
+tar -xf grilo-plugins-0.3.15.tar.xz
+cd grilo-plugins-0.3.15
 mkdir build; cd build
 meson --prefix=/usr --buildtype=minsize ..
 ninja
 ninja install
 install -t /usr/share/licenses/grilo-plugins -Dm644 ../COPYING
 cd ../..
-rm -rf grilo-plugins-0.3.14
+rm -rf grilo-plugins-0.3.15
 # libgtop
 tar -xf libgtop-2.40.0.tar.xz
 cd libgtop-2.40.0