+++ changelog.md
@@ -27,11 +27,13 @@ Upgraded software (GNOME):
 - gssdp: `1.5.1 --> 1.5.2`
 - gupnp: `1.5.2 --> 1.5.3`
 - gtksourceview5: `5.5.0 --> 5.5.1`
+- gtk4: `4.6.6 --> 4.6.7`
+- grilo: `0.3.14 --> 0.3.15`
+- grilo-plugins: `0.3.14 --> 0.3.15`
 - libadwaita: `1.1.3 --> 1.1.4`
 - mutter: `42.3 --> 42.4`
 - rygel: `0.41.1 --> 0.41.2`
 - tracker: `3.3.2 --> 3.3.3`
 - zenity: `3.42.1 --> 3.43.0`
-
 # MassOS 2022.08
 First official release of MassOS with GNOME support.